Part 4 : services and servlets updated to use Pgsql only

+++ datafari-webapp/src/main/java/com/francelabs/datafari/service/db/LicenceDataService.java
@@ -1,39 +1,21 @@
-/*******************************************************************************
- *  * Copyright 2015 France Labs
- *  *
- *  * Licensed under the Apache License, Version 2.0 (the "License");
- *  * you may not use this file except in compliance with the License.
- *  * You may obtain a copy of the License at
- *  *
- *  *      http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
- *******************************************************************************/
 package com.francelabs.datafari.service.db;
 
+import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
-import java.nio.ByteBuffer;
+import java.sql.Blob;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.PreparedStatement;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.apache.zookeeper.server.ByteBufferInputStream;
 
-import com.datastax.oss.driver.api.core.DriverException;
-import com.datastax.oss.driver.api.core.cql.BoundStatement;
-import com.datastax.oss.driver.api.core.cql.PreparedStatement;
-import com.datastax.oss.driver.api.core.cql.ResultSet;
-import com.datastax.oss.driver.api.core.cql.Row;
 import com.francelabs.datafari.exception.CodesReturned;
-import com.francelabs.datafari.exception.DatafariServerException;
 import com.francelabs.licence.Licence;
 
-public class LicenceDataService extends CassandraService {
+public class LicenceDataService {
 
   final static Logger logger = LogManager.getLogger(LicenceDataService.class.getName());
 
@@ -43,18 +25,13 @@ public class LicenceDataService extends CassandraService {
 
   private static LicenceDataService instance;
 
-  public static synchronized LicenceDataService getInstance() throws DatafariServerException {
-    try {
-      if (instance == null) {
-        instance = new LicenceDataService();
-      }
-      instance.refreshSession();
-      return instance;
-    } catch (final DriverException e) {
-      logger.warn("Unable to get instance : " + e.getMessage());
-      // TODO catch specific exception
-      throw new DatafariServerException(CodesReturned.PROBLEMCONNECTIONDATABASE, e.getMessage());
+  private final PostgresService pgService = new PostgresService();
+
+  public static synchronized LicenceDataService getInstance() {
+    if (instance == null) {
+      instance = new LicenceDataService();
     }
+    return instance;
   }
 
   /**
@@ -66,14 +43,16 @@ public class LicenceDataService extends CassandraService {
   public Licence getLicence(final String licenceID) {
     Licence licence = null;
     try {
-      final String query = "SELECT " + LICENCECOLUMN + " FROM " + LICENCECOLLECTION + " where " + IDCOLUMN + "='" + licenceID + "'";
-      final ResultSet result = session.execute(query);
-      final Row row = result.one();
-      if (row != null && !row.isNull(LICENCECOLUMN)) {
-        final ByteBuffer bb = row.getByteBuffer(LICENCECOLUMN);
-        final ByteBufferInputStream bbi = new ByteBufferInputStream(bb);
-        final ObjectInputStream ois = new ObjectInputStream(bbi);
-        licence = (Licence) ois.readObject();
+      String sql = "SELECT " + LICENCECOLUMN + " FROM " + LICENCECOLLECTION + " WHERE " + IDCOLUMN + " = ?";
+      try (ResultSet rs = pgService.executeSelect(sql, licenceID)) {
+        if (rs.next()) {
+          Blob blob = rs.getBlob(LICENCECOLUMN);
+          if (blob != null) {
+            try (ObjectInputStream ois = new ObjectInputStream(blob.getBinaryStream())) {
+              licence = (Licence) ois.readObject();
+            }
+          }
+        }
       }
     } catch (final Exception e) {
       logger.error("Unable to retrieve licence " + licenceID, e);
@@ -84,29 +63,25 @@ public class LicenceDataService extends CassandraService {
   /**
    * Save the licence
    *
-   * @param licenceId
-   *          the licenceId
-   * @param licence
-   *          the licence
+   * @param licenceId the licenceId
+   * @param licence   the licence
    * @return CodesReturned.ALLOK if all was ok
-   * @throws DatafariServerException
    */
   public int saveLicence(final String licenceId, final Licence licence) {
     try {
-      final String existingLicence = getLicence();
+      final String existingLicence = getLicenceId();
       final ByteArrayOutputStream baos = new ByteArrayOutputStream();
       final ObjectOutputStream oos = new ObjectOutputStream(baos);
       oos.writeObject(licence);
-      baos.flush();
-      final ByteBuffer bb = ByteBuffer.wrap(baos.toByteArray());
-      final PreparedStatement ps = session.prepare("INSERT INTO " + LICENCECOLUMN 
-          + " (" + IDCOLUMN + "," 
-          + LICENCECOLUMN + ")" 
-          + " values (?,?)");
-      final BoundStatement bs = ps.bind(licenceId, bb);
-      session.execute(bs);
+      oos.flush();
+      byte[] licenceBytes = baos.toByteArray();
 
-      if (existingLicence != null && !existingLicence.isEmpty()) {
+      // Upsert
+      String sql = "INSERT INTO " + LICENCECOLLECTION + " (" + IDCOLUMN + ", " + LICENCECOLUMN + ") VALUES (?, ?) "
+                 + "ON CONFLICT (" + IDCOLUMN + ") DO UPDATE SET " + LICENCECOLUMN + " = EXCLUDED." + LICENCECOLUMN;
+      pgService.executeUpdate(sql, licenceId, licenceBytes);
+
+      if (existingLicence != null && !existingLicence.isEmpty() && !existingLicence.equals(licenceId)) {
         deleteLicence(existingLicence);
       }
     } catch (final Exception e) {
@@ -115,14 +90,14 @@ public class LicenceDataService extends CassandraService {
     return CodesReturned.ALLOK.getValue();
   }
 
-  private String getLicence() {
+  // Helper to get the current licence id (if exists)
+  private String getLicenceId() {
     try {
-      final String query = "SELECT " + IDCOLUMN + " FROM " + LICENCECOLLECTION + ";";
-      final ResultSet result = session.execute(query);
-      final Row row = result.one();
-      if (row != null && !row.isNull(IDCOLUMN)) {
-        final String licenceId = row.getString(IDCOLUMN);
-        return licenceId;
+      String sql = "SELECT " + IDCOLUMN + " FROM " + LICENCECOLLECTION + " LIMIT 1";
+      try (ResultSet rs = pgService.executeSelect(sql)) {
+        if (rs.next()) {
+          return rs.getString(IDCOLUMN);
+        }
       }
     } catch (final Exception e) {
       logger.warn("Unable to get the licence", e);
@@ -132,10 +107,8 @@ public class LicenceDataService extends CassandraService {
 
   private void deleteLicence(final String licenceId) {
     try {
-      final String query = "DELETE FROM " + LICENCECOLLECTION 
-          + " WHERE " + IDCOLUMN + " = '" + licenceId + "'"
-          + " IF EXISTS";
-      session.execute(query);
+      String sql = "DELETE FROM " + LICENCECOLLECTION + " WHERE " + IDCOLUMN + " = ?";
+      pgService.executeUpdate(sql, licenceId);
     } catch (final Exception e) {
       logger.warn("Unable to delete licence: " + licenceId, e);
     }
@@ -143,11 +116,10 @@ public class LicenceDataService extends CassandraService {
 
   private void emptyLicence() {
     try {
-      final String query = "TRUNCATE " + LICENCECOLLECTION + ";";
-      session.execute(query);
+      String sql = "TRUNCATE TABLE " + LICENCECOLLECTION;
+      pgService.executeUpdate(sql);
     } catch (final Exception e) {
       logger.warn("Unable to empty licence table", e);
     }
   }
-
 }
